Creating NBV_Planner class
